Version v09_37_01_02, patch release for SBN2021C

+++ sbncode/EventGenerator/MeVPrtl/Tools/NuMiKaonGen_tool.cc
@@ -69,7 +69,7 @@ public:
     double LoadPOT();
 
     // no weights
-    float MaxWeight() override { return -1.; }
+    double MaxWeight() override { return -1.; }
 
 private:
   // config
